+++ CHANGELOG.md
@@ -188,3 +188,4 @@ immediately.
 - Switch back from `flaggy` to `cobra`. The latter has fixed the issues that we
   had in the past, and can generate shell completions.
 - Shell completions are now included for `bash`, `fish` and `zsh`.
+- Add a `check` command to verify if a configuration file is valid.